requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
fpdf2==2.7.6
gunicorn==21.2.0
